don't put data in c module anymore
As I almost always import config as c,
I tried to put var in c.var so that I don't need to put them explicitly
in function inputs. However, this is maybe not safe, depending on how
python handles existing data in module load. So, I decided to do things
the old-fashioned way, explicit input variable it is.