store pid_owner_obs instead of searching
in serial mode, the pid_owner_obs is from a search in the obs_list to
look for which pid stores the obs, this is silly, when nproc increases
this operation is taking more and more time to complete. Now, just store
pid_owner_obs along with obs_id in obs_list.