add run_script function to call external mpi runs
scripts/perturb.py, assimilate.py, and diag.py are run externally in a
subprocess, called by the control script run_expt.py which is run in
serial.

Introduce a run_script function to call those scripts and handle errors.
Also, detect if mpi4py is available and if not fall back to a serial
run.