upgrade file lock in Comm
The old one works only with shared memory multiprocessing. Upgraded the file lock to use truely rma parallelism, so that it works for all mpi scenarios

The acquire_file_lock function needs a collective call to generate the MPI.Win object correctly, but in update.py sometimes the processor can be idle (no tasks assigned to it), so it will break the synchronization. Adding an explicit allgather for the filenames and initialize the locks properly.