a few bug fixes
in nc_write_var, if recno is specified for a dimension name, it is okay if it is not unlimited, as long as recno is within the size of that dimension. Relaxing the assertion here.

In model.py for topaz.v5, the diagnostic variables needs to be included in filename if branches to work properly.

The ncat variable check is using the wrong name variable, fixed.