monitor log_file for progress
if log_file is specified, slurm.py submitter will check its content, if not growing at all within a certain amount of time, we kill the job and raise error.

With this the model.run can implement a multiple attempt strategy, to retry the run several times. This solves the random-glitching of compute nodes.